+++ cryptofeed/exchange/blockchain.py
@@ -12,6 +12,7 @@ from itertools import product
 from sortedcontainers import SortedDict as sd
 
 from cryptofeed.defines import BID, ASK, BLOCKCHAIN, BUY, L2_BOOK, L3_BOOK, SELL, TICKER, TRADES
+from cryptofeed.exceptions import MissingSequenceNumber
 from cryptofeed.feed import Feed
 from cryptofeed.standards import pair_exchange_to_std, timestamp_normalize
 
@@ -152,8 +153,11 @@ class Blockchain(Feed):
 
     async def message_handler(self, msg: str, timestamp: float):
         msg = json.loads(msg, parse_float=Decimal)
+        print(msg)
         if self.seq_no is not None and msg['seqnum'] != self.seq_no + 1:
-            raise ValueError("Incorrect sequence number. TODO: implement ws restart")
+            LOG.warning("%s: Missing sequence number detected!", self.id)
+            raise MissingSequenceNumber("Missing sequence number, restarting")
+
         self.seq_no = msg['seqnum']
 
         if 'channel' in msg: