+++ cryptofeed/exchange/bybit.py
@@ -10,7 +10,7 @@ from decimal import Decimal
 from sortedcontainers import SortedDict as sd
 from yapic import json
 
-from cryptofeed.defines import BID, ASK, BUY, BYBIT, L2_BOOK, SELL, TRADES
+from cryptofeed.defines import BID, ASK, BUY, BYBIT, L2_BOOK, SELL, TRADES, OPEN_INTEREST, FUTURES_INDEX
 from cryptofeed.feed import Feed
 from cryptofeed.standards import pair_exchange_to_std as normalize_pair
 from cryptofeed.standards import timestamp_normalize
@@ -40,6 +40,8 @@ class Bybit(Feed):
             await self._trade(msg, timestamp)
         elif "orderBookL2" in msg["topic"]:
             await self._book(msg, timestamp)
+        elif "instrument_info" in msg["topic"]:
+            await self._instrument_info(msg, timestamp)
         else:
             LOG.warning("%s: Invalid message type %s", self.id, msg)
 
@@ -54,6 +56,96 @@ class Bybit(Feed):
                         "args": [f"{chan}.{pair}"]
                     }
                 ))
+
+    async def _instrument_info(self, msg: dict, timestamp: float):
+        """
+        ### Snapshot type update
+        {
+        "topic": "instrument_info.100ms.BTCUSD",
+        "type": "snapshot",
+        "data": {
+            "id": 1,
+            "symbol": "BTCUSD",                           //instrument name
+            "last_price_e4": 81165000,                    //the latest price
+            "last_tick_direction": "ZeroPlusTick",        //the direction of last tick:PlusTick,ZeroPlusTick,MinusTick,
+                                                          //ZeroMinusTick
+            "prev_price_24h_e4": 81585000,                //the price of prev 24h
+            "price_24h_pcnt_e6": -5148,                   //the current last price percentage change from prev 24h price
+            "high_price_24h_e4": 82900000,                //the highest price of prev 24h
+            "low_price_24h_e4": 79655000,                 //the lowest price of prev 24h
+            "prev_price_1h_e4": 81395000,                 //the price of prev 1h
+            "price_1h_pcnt_e6": -2825,                    //the current last price percentage change from prev 1h price
+            "mark_price_e4": 81178500,                    //mark price
+            "index_price_e4": 81172800,                   //index price
+            "open_interest": 154418471,                   //open interest quantity - Attention, the update is not
+                                                          //immediate - slowest update is 1 minute
+            "open_value_e8": 1997561103030,               //open value quantity - Attention, the update is not
+                                                          //immediate - the slowest update is 1 minute
+            "total_turnover_e8": 2029370141961401,        //total turnover
+            "turnover_24h_e8": 9072939873591,             //24h turnover
+            "total_volume": 175654418740,                 //total volume
+            "volume_24h": 735865248,                      //24h volume
+            "funding_rate_e6": 100,                       //funding rate
+            "predicted_funding_rate_e6": 100,             //predicted funding rate
+            "cross_seq": 1053192577,                      //sequence
+            "created_at": "2018-11-14T16:33:26Z",
+            "updated_at": "2020-01-12T18:25:16Z",
+            "next_funding_time": "2020-01-13T00:00:00Z",  //next funding time
+                                                          //the rest time to settle funding fee
+            "countdown_hour": 6                           //the remaining time to settle the funding fee
+        },
+        "cross_seq": 1053192634,
+        "timestamp_e6": 1578853524091081                  //the timestamp when this information was produced
+        }
+
+        ### Delta type update
+        {
+        "topic": "instrument_info.100ms.BTCUSD",
+        "type": "delta",
+        "data": {
+            "delete": [],
+            "update": [
+                {
+                    "id": 1,
+                    "symbol": "BTCUSD",
+                    "prev_price_24h_e4": 81565000,
+                    "price_24h_pcnt_e6": -4904,
+                    "open_value_e8": 2000479681106,
+                    "total_turnover_e8": 2029370495672976,
+                    "turnover_24h_e8": 9066215468687,
+                    "volume_24h": 735316391,
+                    "cross_seq": 1053192657,
+                    "created_at": "2018-11-14T16:33:26Z",
+                    "updated_at": "2020-01-12T18:25:25Z"
+                }
+            ],
+            "insert": []
+        },
+        "cross_seq": 1053192657,
+        "timestamp_e6": 1578853525691123
+        }
+        """
+        update_type = msg['type']
+
+        if update_type == 'snapshot':
+            updates = [msg['data']]
+        else:
+            updates = msg['data']['update']
+
+        for info in updates:
+            if 'open_interest' in info:
+                await self.callback(OPEN_INTEREST, feed=self.id,
+                                    pair=normalize_pair(info['symbol']),
+                                    open_interest=info['open_interest'],
+                                    timestamp=timestamp_normalize(self.id, info['updated_at']),
+                                    receipt_timestamp=timestamp)
+
+            if 'index_price_e4' in info:
+                await self.callback(FUTURES_INDEX, feed=self.id,
+                                    pair=normalize_pair(info['symbol']),
+                                    futures_index=info['index_price_e4'] * 1e-4,
+                                    timestamp=timestamp_normalize(self.id, info['updated_at']),
+                                    receipt_timestamp=timestamp)
 
     async def _trade(self, msg: dict, timestamp: float):
         """