+++ cryptofeed/standards.py
@@ -18,7 +18,7 @@ from cryptofeed.defines import (BINANCE, BINANCE_FUTURES, BINANCE_DELIVERY, BINA
                                 FTX_US, FUNDING, GATEIO, GEMINI, HITBTC, HUOBI, HUOBI_DM, HUOBI_SWAP, IMMEDIATE_OR_CANCEL, KRAKEN,
                                 KRAKEN_FUTURES, L2_BOOK, L3_BOOK, LIMIT, LIQUIDATIONS,
                                 MAKER_OR_CANCEL, MARKET, OKCOIN, OKEX, OPEN_INTEREST, POLONIEX, PROBIT, TICKER,
-                                TRADES, UNSUPPORTED, UPBIT, VOLUME)
+                                TRADES, UNSUPPORTED, UPBIT, VOLUME, FUTURES_INDEX)
 from cryptofeed.exceptions import UnsupportedDataFeed, UnsupportedTradingOption, UnsupportedTradingPair
 from cryptofeed.pairs import gen_pairs, _exchange_info
 
@@ -230,6 +230,7 @@ _feed_to_exchange_map = {
         FTX: 'open_interest',
         BINANCE_FUTURES: 'open_interest',
         BINANCE_DELIVERY: 'open_interest',
+        BYBIT: 'instrument_info.100ms'
     },
     LIQUIDATIONS: {
         BITMEX: 'liquidation',
@@ -237,6 +238,9 @@ _feed_to_exchange_map = {
         BINANCE_DELIVERY: 'forceOrder',
         FTX: 'trades',
         DERIBIT: 'trades'
+    },
+    FUTURES_INDEX: {
+        BYBIT: 'instrument_info.100ms'
     }
 }